# swagger/tindeer.yaml
openapi: 3.0.0
info:
  title: Tindeer API
  description: Notre super API pour notre application mobile de rencontre de Rennes
  version: 1.0.0
servers:
  - url: http://localhost:8000/
    description: Brindeer App
components:
     schemas:
       ProfileDto:
         type: object
         properties:
           id:
             type: string
           userId:
             type: string
           mail:
             type: string
           age:
             type: integer
           firstname:
             type: string
           lastname:
             type: string
           created:
             type: integer
           updated:
             type: integer
...
security:
    - ApiKeyAuth: [ ]
    - oAuthSample:
        - profile
        - email
paths:
  /profile/profile_search:
    get:
      summary: Retourne la liste des profils
      parameters:
        - in: query
          name: mail
          schema:
            type: string
          allowReserved: true
          description: Recherche le profil par mail
      responses:
        '200':
          description: A JSON array of profiles
          content:
            application/json:
              schema:
                type: object
...
  /profile/update:
              put:
                summary: Mettre à jour le profil de l'utilisateur
                requestBody:
                  content:
                    application/json:
                      schema:
                        $ref: "#/components/schemas/ProfileDto"
                responses:
                  '200': # status code
                    description: A JSON array of user names
                    content:
                      application/json:
                        schema:
                          type: object
                          properties:
                            name:
                              type: string
...